Move functions related to organization mapping to their own hook.
